third_party/blink/web_tests/virtual/threaded/external/wpt/css/css-view-transitions/style-inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] style inheritance of pseudo elements
  assert_equals: wrapper expected "0.05s" but got "0s"
Harness: the test ran to completion.
